pyproject.toml
[project]
name = "fitness-tracker"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.17.2",
  "argon2-cffi>=25.1.0",
  "email-validator>=2.3.0",
  "fastapi>=0.110",
  "psycopg[binary]>=3.3.2",
  "pydantic-settings>=2.12.0",
  "pyjwt[crypto]>=2.10.1",
  "sqlalchemy>=2.0.45",
  "uvicorn[standard]>=0.30",
]

[tool.ruff]
line-length = 100
target-version = "py312"
